refactor : 테스트 config 어노테이션 변경
