+++ src/test/java/roomescape/config/TestPaymentConfig.java
@@ -1,7 +1,8 @@
-package roomescape;
+package roomescape.config;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 import roomescape.service.FakePayment;
 import roomescape.service.PaymentClient;
